Recommender

+++ app/backend/Vidb Games/Vidb Games/Models/Entities/Review.cs
@@ -13,8 +13,8 @@ namespace Vidb_Games.Models.Entities
         public string? Comment { get; set; } = string.Empty;
 
         [Required]
-        public int GameId { get; set; }
-        
+        public long GameId { get; set; }
+
         public Game? Game { get; set; }
 
         public int Likes { get; set; } = 0;
@@ -22,7 +22,7 @@ namespace Vidb_Games.Models.Entities
         public int Dislikes { get; set; } = 0;
 
         public Guid UserId { get; set; }
-        
+
         public User? User { get; set; }
     }
 }
